+++ lib/src/config_builder/config_builder.dart
@@ -1,10 +1,12 @@
+import 'package:dart_code_metrics/src/analyzers/unused_files_analyzer/unused_files_config.dart';
+
 import '../analyzers/lint_analyzer/anti_patterns/patterns_factory.dart';
 import '../analyzers/lint_analyzer/lint_config.dart';
 import '../analyzers/lint_analyzer/metrics/metrics_factory.dart';
 import '../analyzers/lint_analyzer/metrics/models/metric.dart';
 import '../analyzers/lint_analyzer/models/entity_type.dart';
 import '../analyzers/lint_analyzer/rules/rules_factory.dart';
-import '../cli/arguments_builder/models/parsed_arguments.dart';
+import '../cli/models/parsed_arguments.dart';
 import '../utils/exclude_utils.dart';
 import 'models/analysis_options.dart';
 import 'models/config.dart';
@@ -54,4 +56,16 @@ class ConfigBuilder {
         prepareExcludes(config.excludeForMetricsPatterns, rootPath),
         config.metrics,
       );
+
+  static UnusedFilesConfig getUnusedFilesConfig(
+    String rootPath,
+    Iterable<String> excludePatterns,
+  ) =>
+      UnusedFilesConfig(prepareExcludes(
+        [
+          ..._defaultSkippedFolders,
+          ...excludePatterns,
+        ],
+        rootPath,
+      ));
 }