+++ lib/src/config_builder/models/config.dart
@@ -1,7 +1,7 @@
 import 'package:meta/meta.dart';
 
 import '../../analyzers/lint_analyzer/metrics/metrics_factory.dart';
-import '../../cli/arguments_builder/models/parsed_arguments.dart';
+import '../../cli/models/parsed_arguments.dart';
 import '../analysis_options_utils.dart';
 import 'analysis_options.dart';