+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add static code diagnostics `always-remove-listener` and `avoid-unnecessary-setstate`.
 * Remove deprecated `Lines of Executable Code` metric, use `Source lines of Code` instead.
 * Changed the supported `analyzer` version to `^1.7.0`.
+* Introduce `analyze` and `check-unused-files` commands.
 
 ## 3.3.4